Looked at filtering out clusters and combining clusters based on proximity and gradient values
To combine clusters, currently we have that they need to be closeby, and that their parameters for the fitted line need to be consistent. To filter out clusters, we currently have that the % error needs to be less than a specified value for it do be defined as a spiral arm part.