Add halo_metrics.ipynb for halo analysis
Currently, planning to differentiate between binned images of dm halo and baryonic gas. Next steps are to figure out a way to run a clustering algorithm (or any algorithm) to find a structure for each image, differentiate to get residuals and then compare with original structures to see any results.